Use separate instances for server and client replicator
