+++ src/client/driver/command-executors/execute-client-code/replicator.js
@@ -1,11 +1,12 @@
-import { identity } from 'lodash';
 import Replicator from 'replicator';
+
+var identityFn = val => val;
 
 // NOTE: we will serialize replicator results
 // to JSON with a command or command result.
 // Therefore there is no need to do additional job here,
 // so we use identity functions for serialization.
 export default new Replicator({
-    serialize:   identity,
-    deserialize: identity
+    serialize:   identityFn,
+    deserialize: identityFn
 });
